% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votedates.R
\name{available_votedates}
\alias{available_votedates}
\title{Get a vector of available vote dates}
\usage{
available_votedates(geolevel = "national")
}
\arguments{
\item{geolevel}{geographical level for which available votedates should be displayed. options "national" or "cantonal"}
}
\value{
a vector of votedates (Format: YYYY-MM-DD)
}
\description{
\code{available_votedates} is a utility function to get the available votedates.
}
\details{
get_swissvotes - retrieve real time vote results for national ballots at district- or municipality level.
}
\examples{
 \donttest{

# Get vector of all available dates
federal_votedates <- available_votedates()

cantonal_votedates <- available_votedates(geolevel="canton")

}

}
