% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_dfr.R
\name{canton_json_to_dfr}
\alias{canton_json_to_dfr}
\title{Transform cantonal results-json into tibble}
\usage{
canton_json_to_dfr(votedate = NULL, geolevel = "municipality",
  dataurl = NULL)
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available.}

\item{geolevel}{geographical level for which the results should be loaded. options."canton","district" or "municipality"}

\item{dataurl}{list of datasets / metadata for the given dataset and its resources OR url of the dcat dataset on opendata.swiss}
}
\value{
a tibble containing the results
}
\description{
\code{canton_json_to_dfr} Tranforms a single results json for a selected cantonal votedata into a tibble.
}
\details{
placeholder
}
\examples{
 \donttest{

#most recent vote
results <- canton_json_to_dfr()

glimpse(results)


}

}
