% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cantonalelections.R
\name{get_cantonalelections}
\alias{get_cantonalelections}
\title{Get Cantonal Election Results}
\usage{
get_cantonalelections(geolevel = "canton")
}
\arguments{
\item{geolevel}{geographical level for which the results should be loaded. Options: "national", "canton", "district", "municipality" or "zh_counting_districts".}
}
\value{
tibble
}
\description{
\code{get_cantonalvotes} Retrieves cantonal election results
}
\examples{
# get cantonal level results
get_cantonalelection()

# get cantonal level results
get_cantonalelection(geolevel="district")
}
