% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_cantonalvotes}
\alias{get_cantonalvotes}
\title{Get cantonal results and counting status}
\usage{
get_cantonalvotes(votedate = NULL, geolevel = "municipality")
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available.}

\item{geolevel}{geographical level for which the results should be loaded. options."canton","district" or "municipality"}
}
\value{
a tibble containing the results
}
\description{
\code{get_cantonalvotes} is one of the two main functions of swissvote package. It allows to retrieve the results and the counting status for national ballots.
}
\details{
get_swissvotes - retrieve real time vote results for cantonal ballots at district- or municipality level.

placeholder
}
\examples{
 \donttest{
results <- get_cantonalvotes(votedate="20181125",geolevel = "municipality")

glimpse(results)


}

}
