% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_stream.R
\name{get_cantonalvotes_stream}
\alias{get_cantonalvotes_stream}
\title{Get cantonal results and counting status in real time}
\usage{
get_cantonalvotes_stream(votedate = NULL, geolevel = "municipality")
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available.}

\item{geolevel}{geographical level for which the results should be loaded. options."canton","district" or "municipality"}
}
\value{
a tibble containing the results
}
\description{
\code{get_cantonalvotes_stream} is one of the two main functions of swissvote package. It allows to retrieve the results and the counting status for national ballots on vote sundays.
}
\details{
get_cantonalvotes_stream - retrieve real time vote results for cantonal ballots at district- or municipality level.

placeholder
}
\examples{
 \donttest{
results <- get_cantonalvotes_stream(votedate="20191002",geolevel = "municipality")

glimpse(results)


}

}
