% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poll.R
\name{get_poll}
\alias{get_poll}
\title{Download poll data collected after a national vote by gfs.bern and the political science departements of the universities of Berne, Zurich, and Geneva.}
\usage{
get_poll(bfsnr = NULL, codebook = F)
}
\arguments{
\item{bfsnr}{number of identification of the vote, by default = NULL. Polls available after September 2020 (from voteid 6360 onwards). Bfsnr corresponds to anr in swissvotes data and has to be four digits (available through get_swissvotes).}

\item{codebook}{by default = FALSE. If TRUE navigates your browser to the codebook if available.}
}
\value{
a tibble containing the results
}
\description{
\code{get_poll} downloads exit poll data Please cite data.
}
\details{
get_poll - retrieve poll data on votes. The unit of analysis are individuals.
}
\examples{
results <- get_poll(bfsnr=6360, codebook=FALSE)
}
