% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_swissvotes}
\alias{get_swissvotes}
\title{Get national results and counting status}
\usage{
get_swissvotes(votedate = NULL, geolevel = "municipality")
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available.}

\item{geolevel}{geographical level for which the results should be loaded. options "district" or "municipality"}
}
\value{
a tibble containing the results
}
\description{
\code{get_swissvotes} is one of the two main functions of swissvote package. It allows to retrieve the results and the counting status for national ballots.
}
\details{
get_swissvotes - retrieve real time vote results for national ballots at district- or municipality level.

placeholder
}
\examples{
 \donttest{
results <- get_swissvotes(votedate="20191002", geolevel = "district")

glimpse(results)


}

}
