% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_range.R
\name{get_swissvotes_range}
\alias{get_swissvotes_range}
\title{Get national results and counting status}
\usage{
get_swissvotes_range(geolevel = "municipality", from_date = NULL,
  to_date = NULL)
}
\arguments{
\item{geolevel}{geographical level for which the results should be loaded. options "district" or "municipality"}

\item{votedate}{date of the ballot. Default: most recent ballot available. Format = YYYYMMDD}
}
\value{
a tibble containing the results
}
\description{
\code{get_swissvotes_range} allows to select a date range to retrieve all the results of the votes that have taken place during the selected period.
}
\details{
get_swissvotes_range - retrieve vote results for national ballots at district- or municipality level for a given date range.

placeholder
}
\examples{
 \donttest{
results <-get_swissvotes_range(geolevel="district",from_date = 20180101,to_date=20181231)

glimpse(results)


}

}
