% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkeys-data.R
\docType{data}
\name{mkeys}
\alias{mkeys}
\title{Merge data}
\format{
An object of class \code{data.frame} with 297 rows and 3 columns.
}
\source{
Walder, Maxime ([Twitter](https://twitter.com/WalderMaxime))
}
\usage{
data(mkeys)
}
\description{
Merge keys that allow to easily combine aggregate data from 
swissvotes with post-ballot surveys (VOX).
}
\examples{
data(mkeys)
bsnr <- mkeys$bfsnr
voxnr <- mkeys$projetx
}
\keyword{datasets}
