% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_dfr.R
\name{swiss_json_to_dfr}
\alias{swiss_json_to_dfr}
\title{Transform swiss results-json into tibble}
\usage{
swiss_json_to_dfr(votedate = NULL, geolevel = "municipality",
  dataurl = NULL)
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available. To select multiple ballots use the 'get_swissvotes'-function. Format = YYYYMMDD}

\item{geolevel}{geographical level for which the results should be loaded. options "national", "canton", "district" or "municipality"}

\item{dataurl}{url of the dataset on opendata-swiss}
}
\value{
a tibble containing the results
}
\description{
\code{swiss_json_to_dfr} Tranforms a .json containing the results of a selected federal votedate in an tibble.
}
\examples{
 \donttest{

# get most recent vote
results <- swiss_json_to_dfr()

glimpse(results)

#get a selected votedate

swiss_json_to_dfr(votedate = "2019-02-10")


}
}
