% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votedates.R
\name{available_votedates}
\alias{available_votedates}
\title{Get a vector of available vote dates}
\usage{
available_votedates(list)
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available. Format = YYYYMMDD}

\item{geolevel}{geographical level for which the results should be loaded. options "district" or "municipality"}
}
\value{
a vector of votedates (Format: YYYYMMDD)
}
\description{
\code{available_votedates} is a utility function to get the available votedates.
}
\details{
get_swissvotes - retrieve real time vote results for national ballots at district- or municipality level.

placeholder
}
\examples{
 \donttest{

# Get data for four votedates 
a <- available_votedates()
votedata <- map_dfr(a[1:4],~get_swissvotes(votedate = .x))


}

}
