% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votedates.R
\name{available_votedates}
\alias{available_votedates}
\title{Get a vector of available vote dates}
\usage{
available_votedates(geolevel = "national")
}
\arguments{
\item{votedate}{date of the ballot. Default: most recent ballot available. Format = YYYYMMDD}

\item{geoevel}{geographical level for which available votedates should be displayed. options "national" or "cantonal"}
}
\value{
a vector of votedates (Format: YYYY-MM-DD)
}
\description{
\code{available_votedates} is a utility function to get the available votedates.
}
\details{
get_swissvotes - retrieve real time vote results for national ballots at district- or municipality level.

placeholder
}
\examples{
 \donttest{

# Get vector of all available dates
all <- available_votedates()



}

}
