% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_votedates.R
\name{available_votedates}
\alias{available_votedates}
\title{Get a vector of available vote dates via `get_nationalvotes` and `get_cantonalvotes``}
\usage{
available_votedates(geolevel = "national", call_res)
}
\arguments{
\item{geolevel}{geographical level for which available votedates should be displayed. options "national" or "canton".}

\item{call_res}{result of a previous call to the base API. Optional argument.}
}
\value{
a vector of votedates (Format: YYYY-MM-DD)
}
\description{
\code{available_votedates} is a utility function to get the available votedates.
}
\details{
available_votedates - get available votedates of federal and cantonal popular votes
}
\examples{

# Get vector of all available dates
federal_votedates <- available_votedates()
cantonal_votedates <- available_votedates(geolevel = "canton")

}
